old_v1/requirements.txt
# === Core Utilities ===
python-dotenv
tqdm
pandas
tabulate

# === OpenAI, Anthropic, Gemini APIs ===
openai>=1.0.0
anthropic
google-generativeai

# === HuggingFace Transformers & Download Helpers ===
transformers>=4.39.0
accelerate
sentencepiece
sacremoses
protobuf
bitsandbytes

# === PyTorch (for model inference) ===
torch
torchvision
torchaudio

# === Sentence Embeddings ===
sentence-transformers

# === Tokenizer Utilities ===
tiktoken

# === Optional but recommended ===
numpy
scipy
